+++ tests/src/sensemaker_dna/sensemaker/neighbourhood.ts
@@ -10,8 +10,53 @@ import {
 import { decode } from "@msgpack/msgpack";
 import { ok } from "assert";
 import pkg from "tape-promise/tape";
-import { installAgent, setUpAliceandBob } from "../../utils";
+import { installAgent } from "../../utils";
 const { test } = pkg;
+
+export const setUpAliceandBob = async (
+  with_config: boolean = false,
+  resource_base_type?: any
+) => {
+  const alice = await createConductor();
+  const bob = await createConductor();
+  const {
+    agentsHapps: alice_happs,
+    agent_key: alice_agent_key,
+    ss_cell_id: ss_cell_id_alice,
+    provider_cell_id: provider_cell_id_alice,
+  } = await installAgent(
+    alice,
+    "alice",
+    undefined,
+    with_config,
+    resource_base_type
+  );
+  const {
+    agentsHapps: bob_happs,
+    agent_key: bob_agent_key,
+    ss_cell_id: ss_cell_id_bob,
+    provider_cell_id: provider_cell_id_bob,
+  } = await installAgent(
+    bob,
+    "bob",
+    alice_agent_key,
+    with_config,
+    resource_base_type
+  );
+  await addAllAgentsToAllConductors([alice, bob]);
+  return {
+    alice,
+    bob,
+    alice_happs,
+    bob_happs,
+    alice_agent_key,
+    bob_agent_key,
+    ss_cell_id_alice,
+    ss_cell_id_bob,
+    provider_cell_id_alice,
+    provider_cell_id_bob,
+  };
+};
 
 export default () => {
   test("SM entry type CRUD tests", async (t) => {
@@ -27,7 +72,7 @@ export default () => {
         ss_cell_id_bob,
         provider_cell_id_alice,
         provider_cell_id_bob,
-      } = await setUpAliceandBob(scenario);
+      } = await setUpAliceandBob();
 
       const callZomeAlice = async (
         zome_name,
@@ -360,7 +405,7 @@ export default () => {
         ss_cell_id_bob,
         provider_cell_id_alice,
         provider_cell_id_bob,
-      } = await setUpAliceandBob(scenario);
+      } = await setUpAliceandBob();
 
       const callZomeAlice = async (
         zome_name,
